# opt: remove interface-to-interface assertion