This commit removes a type assertion from `opt.Expr` to
`opt.ScalarExpr`, which is fairly expensive. Instead, the cheaper
`opt.IsScalarOp` function is used.

Epic: None

Release note: None