@mgartner 2023-11-02
I noticed the interface-to-interface assertions eating up a bit of CPU when profiling some optimizer benchmarks. The gains from this change are minimal (at best -1.8%), but the change is trivial, so I figured "why not?".

```
name                         old time/op    new time/op    delta
SlowQueries/slow-query-1-24    28.6ms ± 1%    28.5ms ± 1%    ~     (p=0.421 n=5+5)
SlowQueries/slow-query-2-24     413ms ± 0%     408ms ± 1%  -1.36%  (p=0.008 n=5+5)
SlowQueries/slow-query-3-24    96.9ms ± 0%    97.5ms ± 0%  +0.68%  (p=0.016 n=5+5)
SlowQueries/slow-query-4-24     691ms ± 0%     694ms ± 1%    ~     (p=0.151 n=5+5)

name                         old alloc/op   new alloc/op   delta
SlowQueries/slow-query-1-24    6.83MB ± 0%    6.83MB ± 0%    ~     (p=0.421 n=5+5)
SlowQueries/slow-query-2-24    48.6MB ± 0%    48.6MB ± 0%    ~     (p=0.548 n=5+5)
SlowQueries/slow-query-3-24    44.8MB ± 0%    44.8MB ± 0%    ~     (p=0.548 n=5+5)
SlowQueries/slow-query-4-24     153MB ± 0%     153MB ± 0%    ~     (p=1.000 n=5+5)

name                         old allocs/op  new allocs/op  delta
SlowQueries/slow-query-1-24     70.2k ± 0%     70.2k ± 0%    ~     (p=0.317 n=5+5)
SlowQueries/slow-query-2-24      391k ± 0%      391k ± 0%    ~     (p=0.516 n=5+5)
SlowQueries/slow-query-3-24      389k ± 0%      389k ± 0%    ~     (p=0.690 n=5+5)
SlowQueries/slow-query-4-24     3.23M ± 0%     3.23M ± 0%    ~     (p=1.000 n=5+5)
```

```
name                                                    old time/op    new time/op    delta
Phases/kv-read/Simple/Explore-24                          51.1µs ± 1%    50.9µs ± 1%    ~     (p=0.151 n=5+5)
Phases/kv-read-const/Simple/Explore-24                    51.8µs ± 1%    51.4µs ± 1%    ~     (p=0.421 n=5+5)
Phases/tpcc-new-order/Simple/Explore-24                    137µs ± 1%     136µs ± 1%    ~     (p=0.151 n=5+5)
Phases/tpcc-delivery/Simple/Explore-24                     116µs ± 0%     115µs ± 1%  -0.83%  (p=0.016 n=4+5)
Phases/tpcc-delivery/Prepared/Explore-24                  43.7µs ± 1%    43.6µs ± 1%    ~     (p=1.000 n=5+5)
Phases/tpcc-stock-level/Simple/Explore-24                  616µs ± 1%     611µs ± 1%    ~     (p=0.056 n=5+5)
Phases/tpcc-stock-level/Prepared/Explore-24                267µs ± 0%     265µs ± 1%    ~     (p=0.151 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24        460µs ± 0%     459µs ± 1%    ~     (p=0.548 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/Explore-24      403µs ± 0%     404µs ± 0%    ~     (p=0.690 n=5+5)
Phases/many-columns-and-indexes-b/Simple/Explore-24        549µs ± 0%     546µs ± 1%  -0.49%  (p=0.032 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/Explore-24      459µs ± 0%     459µs ± 0%    ~     (p=1.000 n=5+5)
Phases/many-columns-and-indexes-c/Simple/Explore-24        342ms ± 1%     336ms ± 1%  -1.81%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24      348ms ± 2%     343ms ± 1%  -1.62%  (p=0.032 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24        139µs ± 0%     138µs ± 0%  -0.77%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Prepared/Explore-24     59.2µs ± 1%    59.1µs ± 0%    ~     (p=0.548 n=5+5)
Phases/batch-insert-one/Simple/Explore-24                  120µs ± 1%     120µs ± 1%    ~     (p=0.841 n=5+5)
Phases/batch-insert-many/Simple/Explore-24                8.69ms ± 1%    8.62ms ± 1%    ~     (p=0.095 n=5+5)
Phases/ored-preds-100/Simple/Explore-24                   3.67ms ± 0%    3.67ms ± 0%    ~     (p=0.841 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-24      3.48ms ± 0%    3.48ms ± 0%    ~     (p=0.151 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24    1.54ms ± 0%    1.54ms ± 0%    ~     (p=0.421 n=5+5)

name                                                    old alloc/op   new alloc/op   delta
Phases/kv-read/Simple/Explore-24                          15.3kB ± 0%    15.3kB ± 0%    ~     (p=0.556 n=5+4)
Phases/kv-read-const/Simple/Explore-24                    15.4kB ± 0%    15.4kB ± 0%    ~     (p=0.444 n=5+5)
Phases/tpcc-new-order/Simple/Explore-24                   32.7kB ± 0%    32.7kB ± 0%    ~     (p=0.659 n=5+5)
Phases/tpcc-delivery/Simple/Explore-24                    26.7kB ± 0%    26.7kB ± 0%    ~     (p=0.619 n=5+5)
Phases/tpcc-delivery/Prepared/Explore-24                  12.8kB ± 0%    12.8kB ± 0%    ~     (p=0.556 n=4+5)
Phases/tpcc-stock-level/Simple/Explore-24                  130kB ± 0%     130kB ± 0%    ~     (p=0.310 n=5+5)
Phases/tpcc-stock-level/Prepared/Explore-24               58.9kB ± 0%    58.9kB ± 0%    ~     (p=0.667 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24       39.6kB ± 0%    39.6kB ± 0%    ~     (p=0.587 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/Explore-24     5.50kB ± 0%    5.50kB ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Simple/Explore-24       47.9kB ± 0%    47.9kB ± 0%    ~     (p=0.968 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/Explore-24     8.02kB ± 0%    8.02kB ± 0%    ~     (p=0.921 n=5+5)
Phases/many-columns-and-indexes-c/Simple/Explore-24        110MB ± 0%     110MB ± 0%    ~     (p=0.690 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24      109MB ± 0%     109MB ± 0%    ~     (p=0.310 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24       29.7kB ± 0%    29.7kB ± 0%    ~     (p=0.810 n=5+5)
Phases/single-col-histogram-range/Prepared/Explore-24     13.0kB ± 0%    13.0kB ± 0%    ~     (all equal)
Phases/batch-insert-one/Simple/Explore-24                 64.8kB ± 0%    64.8kB ± 0%    ~     (p=0.159 n=5+5)
Phases/batch-insert-many/Simple/Explore-24                3.92MB ± 0%    3.92MB ± 0%    ~     (p=0.548 n=5+5)
Phases/ored-preds-100/Simple/Explore-24                    945kB ± 0%     945kB ± 0%    ~     (p=0.548 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-24       926kB ± 0%     926kB ± 0%    ~     (p=0.524 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24     736kB ± 0%     736kB ± 0%    ~     (p=1.000 n=5+5)

name                                                    old allocs/op  new allocs/op  delta
Phases/kv-read/Simple/Explore-24                            95.0 ± 0%      95.0 ± 0%    ~     (all equal)
Phases/kv-read-const/Simple/Explore-24                      95.0 ± 0%      95.0 ± 0%    ~     (all equal)
Phases/tpcc-new-order/Simple/Explore-24                      175 ± 0%       175 ± 0%    ~     (all equal)
Phases/tpcc-delivery/Simple/Explore-24                       165 ± 0%       165 ± 0%    ~     (all equal)
Phases/tpcc-delivery/Prepared/Explore-24                    75.0 ± 0%      75.0 ± 0%    ~     (all equal)
Phases/tpcc-stock-level/Simple/Explore-24                    708 ± 0%       708 ± 0%    ~     (all equal)
Phases/tpcc-stock-level/Prepared/Explore-24                  341 ± 0%       341 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-a/Simple/Explore-24         81.0 ± 0%      81.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-a/Prepared/Explore-24       31.0 ± 0%      31.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Simple/Explore-24          136 ± 0%       136 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Prepared/Explore-24       49.0 ± 0%      49.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-c/Simple/Explore-24         762k ± 0%      762k ± 0%    ~     (p=0.690 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24       756k ± 0%      756k ± 0%    ~     (p=0.595 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24          522 ± 0%       522 ± 0%    ~     (all equal)
Phases/single-col-histogram-range/Prepared/Explore-24        251 ± 0%       251 ± 0%    ~     (all equal)
Phases/batch-insert-one/Simple/Explore-24                    325 ± 0%       325 ± 0%    ~     (all equal)
Phases/batch-insert-many/Simple/Explore-24                 28.5k ± 0%     28.5k ± 0%    ~     (p=0.881 n=5+5)
Phases/ored-preds-100/Simple/Explore-24                    4.53k ± 0%     4.53k ± 0%    ~     (all equal)
Phases/ored-preds-using-params-100/Simple/Explore-24       4.53k ± 0%     4.53k ± 0%    ~     (p=1.000 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24     2.88k ± 0%     2.88k ± 0%    ~     (all equal)
```